chore: adds unit tests, stable
