+++ apps/wjt/src/scripts/blog-post/blog-post.mocks.ts
@@ -1,4 +1,7 @@
 import { DefaultFrontMatter } from './blog-post';
+import { readFileSync } from 'fs';
+import { join } from 'path';
+import { cwd } from 'process';
 
 export const rawPosts = [
   [
@@ -66,3 +69,13 @@ export const rawContentMocks = [
   getMockPostContent(rawPosts[1].slice(6)),
   getMockPostContent(rawPosts[2].slice(5)),
 ];
+
+export const mockFileSystem = {
+  'src/posts/example-post.md': rawPostMocks[0],
+  'src/posts/example-post-2.md': rawPostMocks[1],
+  'src/posts/example-post-3.md': rawPostMocks[2],
+  'src/views/templates/head.pug': readFileSync(
+    join(cwd(), 'src/views/templates/head.pug'),
+    'utf8'
+  ),
+};
